tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-state-overflowing-container-type-change.html

<!DOCTYPE html>
<title>@container: scroll-state(overflowing) property changes</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#overflowing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  #scroller {
    container-type: scroll-state;
    overflow: auto;
    width: 100px;
    height: 100px;
  }
  #target {
    width: 200px;
    height: 200px;
    --overflowing: no;
    @container scroll-state(overflowing) {
      --overflowing: yes;
    }
  }
</style>
<div id="scroller">
  <div id="target"></div>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--overflowing"), "yes");
  }, "Check scroll-state(overflowing) initially matching");

  promise_test(async t => {
    t.add_cleanup(async () => scroller.style.overflow = "");
    scroller.style.overflow = "initial";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--overflowing"), "no",
                  "overflow removed");
    scroller.style.overflow = "";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--overflowing"), "yes",
                  "overflow re-added");
  }, "Check scroll-state(overflowing) not matching when overflow is removed");

  promise_test(async t => {
    t.add_cleanup(async () => scroller.style.containerType = "");
    scroller.style.containerType = "initial";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--overflowing"), "no",
                  "container-type removed");
    scroller.style.containerType = "";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--overflowing"), "yes",
                  "container-type re-added");
  }, "Check scroll-state(overflowing) not matching when container-type is removed");

</script>
